agent: relocate validateComponentsConfig + hasTransitiveDep
- validateComponentsConfig + LEGACY_QA_TOP_LEVEL_KEYS → project-config-store.ts
- hasTransitiveDep → new agent/gate-deps.ts (workflow-validator.ts is
  already taken by load-time workflow validation, unrelated)

Net: server.ts -61 lines (8032 → 7971). No behaviour change.
