+++ src/server/server.ts
@@ -56,7 +56,8 @@ import { inlineFileImages } from "./agent/inline-file-images.js";
 import { StaffManager } from "./agent/staff-manager.js";
 import { TriggerEngine } from "./agent/staff-trigger-engine.js";
 import { PreferencesStore } from "./agent/preferences-store.js";
-import { ProjectConfigStore } from "./agent/project-config-store.js";
+import { ProjectConfigStore, validateComponentsConfig, LEGACY_QA_TOP_LEVEL_KEYS } from "./agent/project-config-store.js";
+import { hasTransitiveDep } from "./agent/gate-deps.js";
 import { ToolGroupPolicyStore } from "./agent/tool-group-policy-store.js";
 import { getAllConfigDirectories, removeBuiltinDirectory, resetConfigDirectories } from "./agent/config-directories.js";
 import { checkDockerAvailability, buildSandboxImage, isBuildingImage, ensureImageAgentVersion } from "./agent/sandbox-status.js";
@@ -136,55 +137,6 @@ export {
 
 /** Cached Docker availability result to avoid running `docker info` per session creation */
 let _dockerAvailCache: { available: boolean; error?: string; ts: number } | null = null;
-
-/**
- * The seven legacy top-level QA keys that have moved to per-component
- * `config:` maps. Rejected on PUT and stripped from GET responses as
- * defence in depth (state-migration removes them on boot).
- */
-const LEGACY_QA_TOP_LEVEL_KEYS = [
-	"qa_start_command",
-	"qa_build_command",
-	"qa_health_check",
-	"qa_browser_entry",
-	"qa_env",
-	"qa_max_duration_minutes",
-	"qa_max_scenarios",
-] as const;
-
-/**
- * Validate the per-component `config:` map (post-migration, opaque
- * key→string). Rules mirror the propose_project tool's runtime validator:
- *   - keys must be non-empty strings
- *   - values must be strings
- *   - max 100 entries per component
- *
- * Returns null on success, or a string error message suitable for HTTP 400.
- */
-function validateComponentsConfig(components: unknown): string | null {
-	if (!Array.isArray(components)) return null;
-	for (const c of components) {
-		if (!c || typeof c !== "object") continue;
-		const cfg = (c as { config?: unknown }).config;
-		if (cfg === undefined || cfg === null) continue;
-		if (typeof cfg !== "object" || Array.isArray(cfg)) {
-			return `components[${(c as { name?: unknown }).name ?? "?"}].config: must be an object`;
-		}
-		const entries = Object.entries(cfg as Record<string, unknown>);
-		if (entries.length > 100) {
-			return `components[${(c as { name?: unknown }).name ?? "?"}].config: too many entries (max 100, got ${entries.length})`;
-		}
-		for (const [k, v] of entries) {
-			if (typeof k !== "string" || k.length === 0) {
-				return `components[${(c as { name?: unknown }).name ?? "?"}].config: empty key`;
-			}
-			if (typeof v !== "string") {
-				return `components[${(c as { name?: unknown }).name ?? "?"}].config.${k}: must be string, got ${typeof v}`;
-			}
-		}
-	}
-	return null;
-}
 
 export interface TlsConfig {
 	cert: string;  // path to PEM certificate
@@ -7927,19 +7879,6 @@ async function handleApiRoute(
 	}
 
 	json({ error: "Not found" }, 404);
-}
-
-/** Check if gateId transitively depends on targetId in the workflow DAG */
-function hasTransitiveDep(workflow: import("./agent/workflow-store.js").Workflow, gateId: string, targetId: string, visited = new Set<string>()): boolean {
-	if (visited.has(gateId)) return false;
-	visited.add(gateId);
-	const gate = workflow.gates.find(g => g.id === gateId);
-	if (!gate) return false;
-	for (const dep of gate.dependsOn) {
-		if (dep === targetId) return true;
-		if (hasTransitiveDep(workflow, dep, targetId, visited)) return true;
-	}
-	return false;
 }
 
 function readBody(req: http.IncomingMessage): Promise<any> {
